// utils/types.ts
export type RequestType = "GET" | "POST" | "PUT" | "PATCH" | "DELETE";
export type Status = "200" | "400" | "401" | "403" | "404" | "405" | "500";
export type AsyncFnType = (req: NextApiRequest, res: NextApiResponse) => Promise<any>;